role middleware added

+++ routes/web.php
@@ -49,41 +49,42 @@ Route::middleware('auth')
     ->group(function(){
 
         // user
-        Route::name('user.')->prefix('users')->group(function () {
+        Route::name('user.')
+            
+            ->prefix('users')->group(function () {
             Route::get('/dashboard', [HomeController::class, 'index'])->name('dashboard');
-            Route::get('/', [UserController::class, 'index'])->name('index');
-            Route::get('/{id}/fetch', [UserController::class, 'fetch'])->name('fetch');
+            Route::middleware('role')->get('/', [UserController::class, 'index'])->name('index');
+            Route::middleware('role')->get('/{id}/fetch', [UserController::class, 'fetch'])->name('fetch');
 
-            Route::post('/', [UserController::class, 'create'])->name('create');
-            Route::put('/', [UserController::class, 'update'])->name('update');
-            Route::delete('/{id}', [UserController::class, 'delete'])->name('delete');
-            Route::get('/key-generate', [UserController::class, 'keyGenerate'])->name('key-generate');
+            Route::middleware('role')->post('/', [UserController::class, 'create'])->name('create');
+            Route::middleware('role')->put('/', [UserController::class, 'update'])->name('update');
+            Route::middleware('role')->delete('/{id}', [UserController::class, 'delete'])->name('delete');
+            Route::middleware('role')->get('/key-generate', [UserController::class, 'keyGenerate'])->name('key-generate');
         });
         
         // sender-info
         Route::name('sender-info.')->prefix('sender-info')->group(function () {
             Route::get('/', [SenderInfoController::class, 'index'])->name('index');
-            Route::get('/sender-id-generate', [SenderInfoController::class, 'senderIdGenerate'])->name('sender-id-generate');
-            Route::post('/', [SenderInfoController::class, 'create'])->name('create');
-            Route::put('/', [SenderInfoController::class, 'update'])->name('update');
+            Route::middleware('role')->get('/sender-id-generate', [SenderInfoController::class, 'senderIdGenerate'])->name('sender-id-generate');
+            Route::middleware('role')->post('/', [SenderInfoController::class, 'create'])->name('create');
+            Route::middleware('role')->put('/', [SenderInfoController::class, 'update'])->name('update');
             Route::delete('/{id}', [SenderInfoController::class, 'delete'])->name('delete');
-            Route::get('/{id}/fetch', [SenderInfoController::class, 'fetch'])->name('fetch');
+            Route::middleware('role')->get('/{id}/fetch', [SenderInfoController::class, 'fetch'])->name('fetch');
         });
 
         // balance
         Route::name('balance.')->prefix('balances')->group(function () {
             Route::get('/', [BalanceController::class, 'index'])->name('index');
-            Route::post('/', [BalanceController::class, 'store'])->name('store');
-            Route::put('/', [BalanceController::class, 'update'])->name('update');
-            Route::delete('/{id}', [BalanceController::class, 'delete'])->name('delete');
-
-            Route::get('/fetch/sender-info/{id}/by-user', [BalanceController::class, 'senderInfoByUser'])->name('sender-info.by-user');
-            Route::get('/fetch/{id}/', [BalanceController::class, 'fetch'])->name('fetch');
+            Route::middleware('role')->post('/', [BalanceController::class, 'store'])->name('store');
+            Route::middleware('role')->put('/', [BalanceController::class, 'update'])->name('update');
+            Route::middleware('role')->delete('/{id}', [BalanceController::class, 'delete'])->name('delete');
+            Route::middleware('role')->get('/fetch/sender-info/{id}/by-user', [BalanceController::class, 'senderInfoByUser'])->name('sender-info.by-user');
+            Route::middleware('role')->get('/fetch/{id}/', [BalanceController::class, 'fetch'])->name('fetch');
 
         });
         
         // fund
-        Route::name('fund.')->prefix('fund')->group(function () {
+        Route::middleware('role')->name('fund.')->prefix('fund')->group(function () {
             Route::get('/', [FundController::class, 'index'])->name('index');
             Route::post('/', [FundController::class, 'store'])->name('store');
             Route::put('/', [FundController::class, 'update'])->name('update');
@@ -92,7 +93,7 @@ Route::middleware('auth')
         });
 
         // credit
-        Route::name('credit.')->prefix('credit')->group(function () {
+        Route::middleware('role')->name('credit.')->prefix('credit')->group(function () {
             Route::get('/', [CreditController::class, 'index'])->name('index');
             Route::post('/', [CreditController::class, 'store'])->name('store');
             Route::put('/', [CreditController::class, 'update'])->name('update');
